+++ service-worker.js
@@ -1,4 +1,4 @@
-const CACHE_NAME="fortress-commander-v1.18.4";
+const CACHE_NAME="fortress-commander-v1.18.5";
 const INDEX_URL=new URL("index.html",self.registration.scope).href;
 const APP_SHELL=[
  './index.html',
@@ -25,6 +25,7 @@ const APP_SHELL=[
  './js/siege.js',
  './js/specializations.js',
  './js/stone-buildings.js',
+ './js/troops.js',
  './js/ui.js',
  './js/utils.js',
  './js/villagers.js',